+++ lib/flutter_awesome_logger.dart
@@ -4,13 +4,11 @@ export 'src/core/logging_using_logger.dart' show AwesomeLoggerConfig;
 
 // API logger exports
 export 'src/api_logger/api_log_entry.dart';
-export 'src/api_logger/api_logger_service.dart';
 export 'src/api_logger/dio_interceptor.dart';
 
 // Floating logger exports
 export 'src/floating_logger/flutter_awesome_logger.dart';
 export 'src/floating_logger/floating_logger_config.dart';
-export 'src/floating_logger/floating_logger_manager.dart';
 
 // UI exports
 export 'src/ui/logger_history_page.dart';