+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] (2024-10-15)
+
+### 🔧 Fix - Improved API Documentation
+- using `FlutterAwesomeLogger.isVisible()` instead of `FloatingLoggerManager.isVisible()`
+- using `FlutterAwesomeLogger.savePosition()` instead of `FloatingLoggerManager.savePosition()`
+- using `FlutterAwesomeLogger.getSavedPosition()` instead of `FloatingLoggerManager.getSavedPosition()`
+- using `FlutterAwesomeLogger.clearPreferences()` instead of `FloatingLoggerManager.clearPreferences()`
+- using `FlutterAwesomeLogger.initialize()` instead of `FloatingLoggerManager.initialize()`
+
 ## [1.0.0] (2024-10-15)
 
 ### 🎉 Major Release - Stable API